# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Digital Twin Assistant API",
    "description": "Backend API for the Personal Digital Twin productivity assistant",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get Items",
        "description": "Get all items for current user with optional filters",
        "operationId": "get_items_api_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "life_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Life Area"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponse"
                  },
                  "title": "Response Get Items Api Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create Item",
        "description": "Create a new item manually",
        "operationId": "create_item_api_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/history": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get Chat History",
        "description": "Fetch recent chat history for the user",
        "operationId": "get_chat_history_api_chat_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Chat History Api Chat History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Send Message",
        "description": "Main chat endpoint \u2014 routes to the appropriate agent via the Orchestrator.\nReturns a unified response with agent name, message, items, schedule, and/or reflection.",
        "operationId": "send_message_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/schedule": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Get Schedule",
        "description": "Get scheduled items for the user within an optional date range.",
        "operationId": "get_schedule_api_schedule_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date string YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "ISO date string YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date string YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "ISO date string YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Schedule Api Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentChatResponse": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClassifiedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/ScheduleBlock"
            },
            "type": "array",
            "title": "Schedule"
          },
          "reflection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory_candidates": {
            "items": {
              "$ref": "#/components/schemas/MemoryCandidateResponse"
            },
            "type": "array",
            "title": "Memory Candidates"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "message"
        ],
        "title": "AgentChatResponse",
        "description": "Schema for the unified multi-agent chat response"
      },
      "ChatMessage": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatMessage",
        "description": "Schema for incoming chat message"
      },
      "ClassifiedItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "task",
              "idea",
              "thought"
            ],
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "life_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Area"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "priority": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title",
          "category",
          "priority"
        ],
        "title": "ClassifiedItem",
        "description": "Schema for a classified item from AI"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "life_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Area"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "title",
          "category"
        ],
        "title": "ItemCreate",
        "description": "Schema for creating an item manually"
      },
...
      "ItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "life_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Area"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "description",
          "category",
          "subcategory",
          "life_area",
          "deadline",
          "status",
          "priority",
          "created_at",
          "updated_at"
        ],
        "title": "ItemResponse",
        "description": "Schema for item response"
      },
...
      "MemoryCandidateResponse": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "fact": {
            "type": "string",
            "title": "Fact"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.8
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "category",
          "fact"
        ],
        "title": "MemoryCandidateResponse",
        "description": "Unsaved long-term fact candidate that needs user confirmation."
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "agent_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Used"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "timestamp"
        ],
        "title": "MessageResponse",
        "description": "Schema for returning a single chat message from history"
      },
...
      "ReflectionSummary": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Patterns"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "ReflectionSummary",
        "description": "Schema for reflection data from the Reflection/Planner Agent"
      },
      "ScheduleBlock": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "estimated_duration_minutes": {
            "type": "integer",
            "title": "Estimated Duration Minutes"
          },
          "scheduled_start": {
            "type": "string",
            "title": "Scheduled Start"
          },
          "scheduled_end": {
            "type": "string",
            "title": "Scheduled End"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "title",
          "estimated_duration_minutes",
          "scheduled_start",
          "scheduled_end"
        ],
        "title": "ScheduleBlock",
        "description": "Schema for a scheduled time block from the Scheduler Agent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}